Cast flags to short to match struct definition
Without this cast, xlc16 produces the following error:
```
./unix/omrintrospect.c:226:47: error: implicit conversion from
 'int' to 'short' changes value from 57345 to -8191
 [-Werror,-Wconstant-conversion]
        fds[0].events = POLLHUP | POLLERR | POLLNVAL | POLLIN;
                      ~ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~
```

This cast forces the flags to match the field definition

+++ port/unix/omrintrospect.c
@@ -223,7 +223,7 @@ barrier_block_until_poked(barrier_r *barrier, uintptr_t deadline)
 	struct timespec spec;
 
 	fds[0].fd = barrier->descriptor_pair[0];
-	fds[0].events = POLLHUP | POLLERR | POLLNVAL | POLLIN;
+	fds[0].events = (short)(POLLHUP | POLLERR | POLLNVAL | POLLIN);
 	fds[0].revents = 0;
 
 	if (deadline == 0) {
@@ -496,7 +496,7 @@ sem_timedwait_r(sem_t_r *sem, uintptr_t seconds)
 	int deadline = 0;
 	int interval = seconds;
 	fds[0].fd = sem->descriptor_pair[0];
-	fds[0].events = POLLHUP | POLLERR | POLLNVAL | POLLIN;
+	fds[0].events = (short)(POLLHUP | POLLERR | POLLNVAL | POLLIN);
 	fds[0].revents = 0;
 
 	if (seconds == 0) {
